Refactor exception handling and improve DTO usage across services
- Replaced RuntimeException with ResourceNotFoundException in ExpertServiceImpl and PortfolioServiceImpl for better error handling.
- Updated PortfolioServiceImpl and ServicePackageServiceImpl to throw ForbiddenException for unauthorized access instead of ExpertNotFoundException.
- Changed budget type from Double to BigDecimal in JobPostRequest, JobPostResponse, JobSearchRequest, and related repository methods for precision.
- Introduced MessageResponse DTO to avoid exposing sensitive data in Message entity.
- Implemented JWT authentication in WebSocket configuration and added JwtHandshakeChannelInterceptor for secure connections.
- Created UserSummaryResponse DTO for admin user summaries without exposing passwords.
- Added .env.example file for environment variable configuration.
- Refactored Payment and Transaction entities to extend BaseEntity for consistent auditing fields.

+++ aitasker/src/main/java/com/aitasker/expert/dto/request/UpdateExpertProfileRequest.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public class UpdateExpertProfileRequest {
     @NotBlank(message = "Họ và tên không được để trống")
     @Size(min = 2, max = 50, message = "Tên phải từ 2 đến 50 ký tự")
@@ -22,7 +24,7 @@ public class UpdateExpertProfileRequest {
 
     @NotNull(message = "Mức giá theo giờ không được để trống")
     @PositiveOrZero(message = "Mức giá không được là số âm")
-    private Double hourlyRate;
+    private BigDecimal hourlyRate;
 
     // Getters và Setters
     public String getFullName() { return fullName; }
@@ -33,6 +35,6 @@ public class UpdateExpertProfileRequest {
     public void setSkills(String skills) { this.skills = skills; }
     public Integer getExperienceYears() { return experienceYears; }
     public void setExperienceYears(Integer experienceYears) { this.experienceYears = experienceYears; }
-    public Double getHourlyRate() { return hourlyRate; }
-    public void setHourlyRate(Double hourlyRate) { this.hourlyRate = hourlyRate; }
+    public BigDecimal getHourlyRate() { return hourlyRate; }
+    public void setHourlyRate(BigDecimal hourlyRate) { this.hourlyRate = hourlyRate; }
 }
