+++ aitasker/src/main/java/com/aitasker/job/dto/JobPostRequest.java
@@ -3,6 +3,8 @@ package com.aitasker.job.dto;
 import com.aitasker.common.enums.JobStatus;
 import lombok.Setter;
 import lombok.Getter;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -10,7 +12,9 @@ import java.time.LocalDate;
 public class JobPostRequest {
     private String title;
     private String description;
-    private Double budget;
+    // Double -> BigDecimal: khớp kiểu với JobPost.budget, tránh phải convert
+    // qua lại và mất độ chính xác khi tính tiền.
+    private BigDecimal budget;
     private LocalDate deadline;
     private String requiredSkills;
 }
