+++ aitasker/src/main/java/com/aitasker/job/dto/JobSearchRequest.java
@@ -4,12 +4,14 @@ import com.aitasker.common.enums.JobStatus;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class JobSearchRequest {
     private String keyword;
     private String skills;
-    private Double minBudget;
-    private Double maxBudget;
+    private BigDecimal minBudget;
+    private BigDecimal maxBudget;
     private JobStatus status;
 }
